requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
openpyxl>=3.1.0
plotly>=5.18.0
numpy>=1.24.0
scipy>=1.11.0
python-dateutil>=2.8.0
pytest>=7.4.0
